requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning Frameworks
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0
keras>=2.8.0

# Computer Vision & Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0
imageio>=2.9.0

# Medical Imaging Libraries
nibabel>=3.2.0
SimpleITK>=2.1.0
pydicom>=2.2.0
medpy>=0.4.0

# Visualization
matplotlib>=3.5.0
plotly>=5.6.0
seaborn>=0.11.0
vtk>=9.1.0

# Web Framework for UI
streamlit>=1.8.0
gradio>=3.0.0
flask>=2.0.0

# Data Processing & Analysis
jupyterlab>=3.2.0
ipywidgets>=7.6.0
tqdm>=4.62.0

# Model Utilities
tensorboard>=2.8.0
wandb>=0.12.0
mlflow>=1.24.0

# File & System Utilities
pyyaml>=6.0
configparser>=5.2.0
pathlib2>=2.3.0
argparse

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Optional: For advanced visualization
mayavi>=4.7.0
fury>=0.8.0

# Optional: For cloud deployment
boto3>=1.20.0
google-cloud-storage>=2.1.0